Sectors mode, UDMF: Create Brightness Gradient action can now work in "Light" and "Fade" modes. Visual mode: added support for "fogdensity" and "outsidefogdensity" MAPINFO properties. Sector Info panel now shows the number of sector's sidedefs as well as light and fade colors (UDMF only). UDMF: Sector and sidedef flags are now copied/pasted when using "Copy Properties" and "Paste Properties" actions. Fixed: in some cases default texture overrides were used even when corresponding option was disabled. Sector Edit form: light and fade values were not saved when a value was pasted/entered into text box. Sector Edit form: "reset value" button was not shown for light and fade values when selected sectors had different light/fade colors. Reverted Open Map form changes from previous commit (my assumption that map marker is always empty was not correct...) Updated documentation.

+++ branches/GZDoomBuilder/Source/Core/Rendering/Renderer2D.cs
@@ -530,12 +530,12 @@ namespace CodeImp.DoomBuilder.Rendering
 			if(vertexsize < 0) vertexsize = 0;
 			if(vertexsize > 4) vertexsize = 4;
 
-			Matrix scaling = Matrix.Scaling((1f / (float)windowsize.Width) * 2f, (1f / (float)windowsize.Height) * -2f, 1f);
+			Matrix scaling = Matrix.Scaling((1f / windowsize.Width) * 2f, (1f / windowsize.Height) * -2f, 1f);
 			Matrix translate = Matrix.Translation(-(float)windowsize.Width * 0.5f, -(float)windowsize.Height * 0.5f, 0f);
 			graphics.Device.SetTransform(TransformState.View, Matrix.Multiply(translate, scaling));
 			graphics.Device.SetTransform(TransformState.Projection, Matrix.Identity);
 			Vector2D lt = DisplayToMap(new Vector2D(0.0f, 0.0f));
-			Vector2D rb = DisplayToMap(new Vector2D((float)windowsize.Width, (float)windowsize.Height));
+			Vector2D rb = DisplayToMap(new Vector2D(windowsize.Width, windowsize.Height));
 			viewport = new RectangleF(lt.x, lt.y, rb.x - lt.x, rb.y - lt.y);
 			yviewport = new RectangleF(lt.x, rb.y, rb.x - lt.x, lt.y - rb.y);
 		}
